Move RangeCacheRemovalCounters to top level
